requirements.txt
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Federated Learning
flwr>=1.7.0
flwr-datasets>=0.1.0

# Medical Imaging
nibabel>=5.1.0
SimpleITK>=2.3.0

# Data Science & Visualization
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Dataset & API
kaggle>=1.5.16
requests>=2.31.0

# Progress & Logging
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Utilities
PyYAML>=6.0
argparse
pathlib2
pillow>=9.5.0
python-pptx>=0.6.23

# Development
jupyter>=1.0.0
ipykernel>=6.23.0
black>=23.0.0
flake8>=6.0.0


